Se agrega image a country y otras pequeñas modificaciones

+++ src/main/java/com/dev/AdminPopulation/service/NatalityService.java
@@ -10,5 +10,5 @@ public interface NatalityService {
 
 	List<Natality> findAllNatality();
 	
-	List<Natality> natalityByCountry(Long countryId);
+	List<Natality> natalityByCountry(Integer countryId);
 }
