+++ src/main/java/com/dev/AdminPopulation/service/impl/NatalityServiceImpl.java
@@ -26,7 +26,7 @@ public class NatalityServiceImpl implements NatalityService{
 	}
 
 	@Override
-	public List<Natality> natalityByCountry(Long countryId) {
+	public List<Natality> natalityByCountry(Integer countryId) {
 		return (List<Natality>) this.natalityRepository.natalityByCountry(countryId);
 	}
